+++ torchft/process_group.py
@@ -20,7 +20,7 @@ import logging
 import queue
 import threading
 from datetime import timedelta
-from typing import Any, Dict, List, Optional, Tuple, Type, TYPE_CHECKING, Union
+from typing import TYPE_CHECKING, Any, Dict, List, Optional, Tuple, Type, Union
 
 import torch
 import torch.distributed as dist
@@ -31,16 +31,16 @@ import torch.multiprocessing as mp
 from torch.distributed import (
     BroadcastOptions,
     DeviceMesh,
-    get_rank,
-    init_device_mesh,
     PrefixStore,
     ProcessGroup as BaseProcessGroup,
     ProcessGroupGloo as BaseProcessGroupGloo,
     ProcessGroupNCCL as BaseProcessGroupNCCL,
     Store,
     TCPStore,
+    get_rank,
+    init_device_mesh,
 )
-from torch.distributed.distributed_c10d import _world, Work
+from torch.distributed.distributed_c10d import Work, _world
 from torch.futures import Future
 
 if TYPE_CHECKING:
