+++ torchft/process_group_test.py
@@ -10,37 +10,32 @@ import os
 import unittest
 from concurrent.futures import ProcessPoolExecutor, ThreadPoolExecutor
 from datetime import timedelta
-from typing import Any, cast, Dict, Tuple
-from unittest import skipUnless, TestCase
+from typing import Any, Dict, Tuple, cast
+from unittest import TestCase, skipUnless
 from unittest.mock import Mock
 
 import torch
 import torch.distributed as dist
 from torch import nn
 from torch._C._distributed_c10d import (
-    _resolve_process_group,
     AllgatherOptions,
     AllreduceOptions,
     BroadcastOptions,
     ReduceOp,
+    _resolve_process_group,
 )
 from torch.distributed import (
-    _functional_collectives,
-    get_world_size,
     ReduceOp,
     TCPStore,
     Work,
+    _functional_collectives,
+    get_world_size,
 )
 from torch.distributed.device_mesh import init_device_mesh
 
 from torchft.manager import Manager
 from torchft.process_group import (
-    _DummyWork,
-    _ErrorSwallowingWork,
-    _ManagedWork,
     ErrorSwallowingProcessGroupWrapper,
-    extend_device_mesh,
-    ft_init_device_mesh,
     ManagedProcessGroup,
     ProcessGroup,
     ProcessGroupBabyGloo,
@@ -49,6 +44,11 @@ from torchft.process_group import (
     ProcessGroupGloo,
     ProcessGroupNCCL,
     ProcessGroupWrapper,
+    _DummyWork,
+    _ErrorSwallowingWork,
+    _ManagedWork,
+    extend_device_mesh,
+    ft_init_device_mesh,
 )
 
 
@@ -370,64 +370,3 @@ class ProcessGroupTest(TestCase):
         self.assertEqual(manager.report_error.call_count, 0)
         self.assertEqual(manager.wrap_future.call_count, 1)
         self.assertEqual(manager.wait_quorum.call_count, 1)
-
-
-class DeviceMeshTest(TestCase):
-    @staticmethod
-    def _test_init_device_mesh(world_size: int, rank: int) -> None:
-        os.environ["MASTER_ADDR"] = "127.0.0.1"
-        os.environ["MASTER_PORT"] = str(12346)
-        os.environ["RANK"] = str(rank)
-        os.environ["WORLD_SIZE"] = str(4)
-
-        testcase = TestCase()
-
-        manager = Mock(spec=Manager)
-        # Even though we only have 4 workers, we can still initialize (2, 4) mesh.
-        # That's because the replicate group is NOT phystically created in the
-        # real mesh but is virtually added to the mesh via ManagedDeviceMesh.
-        device_mesh = ft_init_device_mesh(
-            device_type="cpu",
-            mesh_shape=(2, world_size),
-            mesh_dim_names=("dp_replicate", "dp_shard"),
-            replicate_dim=0,
-            manager=manager,
-        )
-
-        testcase.assertTrue(
-            isinstance(device_mesh.get_group("dp_replicate"), ManagedProcessGroup)
-        )
-        testcase.assertTrue(
-            not isinstance(device_mesh.get_group("dp_shard"), ManagedProcessGroup)
-        )
-        replicate_group = device_mesh.get_group("dp_replicate")
-        testcase.assertEqual(
-            cast(ManagedProcessGroup, replicate_group)._manager, manager
-        )
-        replicate_mesh = device_mesh["dp_replicate"]
-        testcase.assertEqual(replicate_mesh.get_group(), replicate_group)
-
-        flatten_mesh = device_mesh._flatten("dp")
-        manager.num_participants.return_value = 0
-        testcase.assertEqual(flatten_mesh.size(), world_size)
-        manager.num_participants.return_value = 1
-        testcase.assertEqual(flatten_mesh.size(), world_size)
-        manager.num_participants.return_value = 2
-        testcase.assertEqual(flatten_mesh.size(), world_size * 2)
-
-        testcase.assertEqual(flatten_mesh.get_local_rank(), dist.get_rank())
-
-        device_mesh.get_coordinate()
-        buffer = io.BytesIO()
-        torch.save(device_mesh, buffer)
-        buffer.seek(0)
-        torch.load(buffer, weights_only=False)
-
-    def test_init_device_mesh(self) -> None:
-        with ProcessPoolExecutor(max_workers=4) as executor:
-            futures = []
-            for i in range(4):
-                future = executor.submit(self._test_init_device_mesh, 4, i)
-                futures.append(future)
-            for f in futures:
-                f.result()
